updated length to cast as int

+++ intrinsicArray.py
@@ -2,7 +2,7 @@ import random
 
 def intrinsicArray(endTime, interval):
 
-    length = endTime/interval 
+    length = int(endTime/interval) 
     intrArray = [0] * length
 
     for i in range (0,index):
